requirements.txt
torch
pillow~=10.2.0
PySide6~=6.8.0.1
sympy~=1.13.1
torchvision
tqdm~=4.66.5
numpy~=1.26.3
matplotlib~=3.9.2
python-slugify~=8.0.4